# web/openapi.yaml
openapi: 3.0.3
info:
  title: Taxi Scenario Writer API
  description: |
    タクシールート最適化とシナリオ生成を行うAPIです。

    主な機能:
    - AI テキスト生成（Qwen / Gemini）
    - 住所のジオコーディング
    - ルート最適化
    - AI によるルート自動生成とシナリオ作成

    ## 環境変数
    以下の環境変数が必要です:
    - `GOOGLE_MAPS_API_KEY`: Google Maps API キー（必須）
    - `QWEN_API_KEY`: Qwen API キー（Qwen使用時）
    - `GEMINI_API_KEY`: Gemini API キー（Gemini使用時）
    - `QWEN_REGION`: Qwen のリージョン（`china` または `international`、デフォルト: `international`）

  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: ローカル開発環境
  - url: https://your-production-domain.com
    description: 本番環境
...
paths:
...
  /api/places/geocode:
    post:
      tags:
        - Places & Routes
      summary: 住所のジオコーディング
      description: 住所または場所名から緯度経度を取得します（Google Places API 使用）
      operationId: geocodeAddresses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - addresses
              properties:
                addresses:
                  type: array
                  description: ジオコーディングする住所または場所名のリスト
                  items:
                    type: string
                  example: ["東京駅", "浅草寺", "東京スカイツリー"]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
                    example: true
                  places:
                    type: array
                    items:
                      $ref: '#/components/schemas/GeocodedPlace'
        '400':
          description: リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/routes/optimize:
    post:
      tags:
        - Places & Routes
      summary: ルート最適化
      description: |
        出発地、目的地、経由地点を指定してルートを最適化します（Google Routes API 使用）
        経由地点の順序を最適化し、総距離と所要時間を最小化します
      operationId: optimizeRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - origin
                - destination
                - intermediates
              properties:
                origin:
                  $ref: '#/components/schemas/RouteWaypoint'
                destination:
                  $ref: '#/components/schemas/RouteWaypoint'
                intermediates:
                  type: array
                  description: 経由地点のリスト（順序が最適化されます）
                  items:
                    $ref: '#/components/schemas/RouteWaypoint'
                travelMode:
                  type: string
                  enum: [DRIVE, WALK, BICYCLE, TRANSIT]
                  default: DRIVE
                  description: 移動手段
                optimizeWaypointOrder:
                  type: boolean
                  default: true
                  description: 経由地点の順序を最適化するか
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
                    example: true
                  optimizedRoute:
                    $ref: '#/components/schemas/OptimizedRoute'
        '400':
          description: リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: エラーメッセージ
          example: "GOOGLE_MAPS_API_KEYが設定されていません"
    LatLng:
      type: object
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type: number
          format: double
          example: 35.681236
        longitude:
          type: number
          format: double
          example: 139.767125
    GeocodedPlace:
      type: object
      required:
        - inputAddress
        - formattedAddress
        - location
        - placeId
      properties:
        inputAddress:
          type: string
          description: 入力された住所または場所名
          example: "東京駅"
        formattedAddress:
          type: string
          description: Google が返す正規化された住所
          example: "日本、〒100-0005 東京都千代田区丸の内１丁目"
        location:
          $ref: '#/components/schemas/LatLng'
        placeId:
          type: string
          description: Google Place ID（一意識別子）
          example: "ChIJC3Cf2PuLGGARZ7KXBYZWe5k"
    RouteWaypoint:
      type: object
      properties:
        name:
          type: string
          description: 地点名（識別用、任意）
          example: "東京駅"
        placeId:
          type: string
          description: Google Place ID
          example: "ChIJC3Cf2PuLGGARZ7KXBYZWe5k"
        location:
          $ref: '#/components/schemas/LatLng'
        address:
          type: string
          description: 住所（placeId も location もない場合に使用）
          example: "東京都千代田区丸の内1丁目"
    OptimizedWaypoint:
      type: object
      required:
        - waypoint
        - waypointIndex
      properties:
        waypoint:
          $ref: '#/components/schemas/RouteWaypoint'
        waypointIndex:
          type: integer
          description: 最適化後の順序インデックス
          example: 0
    RouteLeg:
      type: object
      required:
        - distanceMeters
        - durationSeconds
      properties:
        startLocation:
          $ref: '#/components/schemas/LatLng'
        endLocation:
          $ref: '#/components/schemas/LatLng'
        distanceMeters:
          type: integer
          description: 区間の距離（メートル）
          example: 1500
        durationSeconds:
          type: integer
          description: 区間の所要時間（秒）
          example: 300
    OptimizedRoute:
      type: object
      required:
        - orderedWaypoints
        - legs
        - totalDistanceMeters
        - totalDurationSeconds
      properties:
        orderedWaypoints:
          type: array
          description: 最適化された地点の順序
          items:
            $ref: '#/components/schemas/OptimizedWaypoint'
        legs:
          type: array
          description: 各区間の情報
          items:
            $ref: '#/components/schemas/RouteLeg'
        totalDistanceMeters:
          type: integer
          description: 総距離（メートル）
          example: 8500
        totalDurationSeconds:
          type: integer
          description: 総所要時間（秒）
          example: 1800
...